change text filter behavior
originally split query by spaces to get tokens and searched on every token to find matching rows. changed so that query is not split and is searched in its entirety.

if need arises might implement different types of searches that can be set and/or the ability to inject custom search queries.